move response types to shared
