+++ shared/src/schemas/index.ts
@@ -1,2 +1,3 @@
 export * from "./form";
 export * from "./review";
+export * from "./response";
